chore: split off long-running warp-deploy test
